+++ trino/versions.py
@@ -7,13 +7,6 @@ versions = [
         "trino-storage-connector": "451",
     },
     {
-        "product": "455",
-        "java-base": "22",
-        "java-devel": "22",
-        "jmx_exporter": "1.2.0",
-        "trino-storage-connector": "455",
-    },
-    {
         "product": "470",
         "java-base": "23",
         "java-devel": "23",